+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Adds ability to exclude variables from the output using earthdata-varinfo configuration.
 - Adds configuration entries to exclude SMAP L3 string variables.
+- Changes xarray engine from default netcdf4 to h5netcdf.
 
 ## [v1.4.0] - 2025-09-30